fix: api functions
