+++ src/api/config.ts
@@ -1,0 +1,3 @@
+export const MAX_SUGGESTIONS = 3;
+export const MAX_STRING_LENGTH = 90;
+export const MAX_PAGES = 30;
