+++ src/api/index.ts
@@ -1,10 +1,9 @@
-import {
-    getAllPosts,
-    getFeaturePost,
-    getPostBuId,
-    getSimilarPosts,
-} from '@/api/apiUtils';
 import { checkImage } from '@/api/checkImage';
+import { emailSubmit } from '@/api/emailSubmit';
+import { getAllPosts } from '@/api/getAllPosts';
+import { getFeaturePost } from '@/api/getFeaturePost';
+import { getPostBuId } from '@/api/getPostBuId';
+import { getSimilarPosts } from '@/api/getSimilarPosts';
 
 export {
     getSimilarPosts,
@@ -12,4 +11,5 @@ export {
     getPostBuId,
     getFeaturePost,
     checkImage,
+    emailSubmit,
 };
