+++ src/helpers/index.ts
@@ -1,5 +1,4 @@
 import { configUrlParams } from '@/helpers/configUrlParams';
-import { emailSubmit } from '@/helpers/emailSubmit';
 import { extractSearchParams } from '@/helpers/extractSearchParams';
 import { formatDate } from '@/helpers/formatDate';
 import { formatString } from '@/helpers/formatString';
@@ -11,7 +10,6 @@ import { safeFetch } from '@/helpers/safeFetch';
 export {
     formatString,
     safeFetch,
-    emailSubmit,
     getDefaultCategory,
     formatDate,
     extractSearchParams,
